+++ src/db/models/batch.rs
@@ -1,5 +1,6 @@
 use crate::db::RollupStatusType;
 use chrono::NaiveDateTime;
+use rust_decimal::Decimal;
 use serde::Serialize;
 
 #[derive(sqlx::FromRow, Clone, Debug, Serialize)]
@@ -10,7 +11,7 @@ pub struct Batch {
     pub end_chunk_index: i64,
     pub start_block_number: Option<i64>,
     pub end_block_number: Option<i64>,
-    pub total_tx_num: Option<i64>,
+    pub total_tx_num: Option<Decimal>,
     pub rollup_status: RollupStatusType,
     pub commit_tx_hash: Option<String>,
     pub finalize_tx_hash: Option<String>,
